+++ bindings/nodejs/src/secret_manager.rs
@@ -5,7 +5,10 @@ use std::{ops::Deref, sync::Arc};
 
 use iota_sdk_bindings_core::{
     call_secret_manager_method as rust_call_secret_manager_method,
-    iota_sdk::client::secret::{SecretManager, SecretManagerDto},
+    iota_sdk::client::{
+        secret::{SecretManager, SecretManagerDto},
+        stronghold::StrongholdAdapter,
+    },
     Response, Result, SecretManagerMethod,
 };
 use neon::prelude::*;
@@ -97,6 +100,41 @@ pub fn call_secret_manager_method(mut cx: FunctionContext) -> JsResult<JsUndefin
             Ok(())
         });
     });
+
+    Ok(cx.undefined())
+}
+
+pub fn migrate_stronghold_snapshot_v2_to_v3(mut cx: FunctionContext) -> JsResult<JsUndefined> {
+    let current_path = cx.argument::<JsString>(0)?.value(&mut cx);
+    let current_password = cx.argument::<JsString>(1)?.value(&mut cx).into();
+    let salt = cx.argument::<JsString>(2)?.value(&mut cx);
+    let rounds = cx.argument::<JsNumber>(3)?.value(&mut cx);
+    let new_path = cx
+        .argument_opt(4)
+        .map(|opt| opt.downcast_or_throw::<JsString, _>(&mut cx))
+        .transpose()?
+        .map(|opt| opt.value(&mut cx));
+    let new_password = cx
+        .argument_opt(5)
+        .map(|opt| opt.downcast_or_throw::<JsString, _>(&mut cx))
+        .transpose()?
+        .map(|opt| opt.value(&mut cx))
+        .map(Into::into);
+
+    StrongholdAdapter::migrate_snapshot_v2_to_v3(
+        &current_path,
+        current_password,
+        salt,
+        rounds as u32,
+        new_path.as_ref(),
+        new_password,
+    )
+    .or_else(|e| {
+        cx.throw_error(
+            serde_json::to_string(&Response::Error(e.into()))
+                .expect("the response is generated manually, so unwrap is safe."),
+        )
+    })?;
 
     Ok(cx.undefined())
 }